Auth jwt

+++ resources/store.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import Flask, request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from schemas import StoreSchema
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 from db import db
@@ -9,7 +8,7 @@ from models import StoreModel
 
 # Create Blueprint for the stores
 
-blp = Blueprint("stores", __name__, description="Opertations on stores")
+blp = Blueprint("Stores", "stores", description="Opertations on stores")
 
 
 # Endpoints start with /store
@@ -23,6 +22,7 @@ class StoreList(MethodView):
     
 
     # Create new store
+    @jwt_required
     @blp.arguments(StoreSchema)
     @blp.response(200, StoreSchema)
     def post(self, store_data):
@@ -39,7 +39,7 @@ class StoreList(MethodView):
 
 
 # Endpoint start with /store/store_id
-@blp.route("/store/<string:store_id>")
+@blp.route("/store/<int:store_id>")
 class Store(MethodView):
 
     # get a store with specific id
@@ -50,7 +50,11 @@ class Store(MethodView):
 
 
     # delete a specific store with store_id
+    @jwt_required
     def delete(self, store_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         store = StoreModel.query.get_or_404(store_id)
         db.session.delete(store)
         db.session.commit()
